Fix festival cleanup

+++ jyotisha/panchaanga/writer/generation_project/__init__.py
@@ -58,10 +58,13 @@ def dump_detailed(year, city, year_type, computation_system=ComputationSystem.MU
 
 
 def dump_summary(year, city, script=sanscript.DEVANAGARI, year_type=era.ERA_GREGORIAN, computation_system=ComputationSystem.MULTI_NEW_MOON_SIDEREAL_MONTH_ADHIKA_AMAANTA__CHITRA_180, allow_precomputed=False):
+  out_path = get_canonical_path(city=city.name, computation_system_str=str(computation_system), year=year, year_type=year_type)
+  if os.path.exists(out_path + ".toml"):
+    logging.info(f"{out_path}.toml extists. skipping")
+    return 
   logging.info("Generating summary panchaanga for %s year %d (%s), with computation system %s ", city.name, year, year_type, str(computation_system))
   panchaanga = annual.get_panchaanga_for_year(city=city, year=year, computation_system=computation_system, year_type=year_type, allow_precomputed=allow_precomputed)
   year_table = to_table_dict(panchaanga=panchaanga )
-  out_path = get_canonical_path(city=panchaanga.city.name, computation_system_str=str(panchaanga.computation_system), year=year, year_type=year_type)
   os.makedirs(os.path.dirname(out_path), exist_ok=True)
   with codecs.open(out_path + ".toml", "w") as fp:
     toml.dump(year_table, fp)
